fix(users): renamed namespaces
